requirements.txt
torch
torchvision
transformers
PyWavelets
scipy
numpy
pandas
Pillow
tqdm
matplotlib
requests
yfinance
